Create more test helpers
* Add setupIntegrationTests()

* Add assertRequest()

* Add goToSearchPage()

* Handle multiple date of birth fields in fillExpungementPacketForm()
