+++ src/frontend/src/test/testHelpers.tsx
@@ -1,24 +1,36 @@
 // https://redux.js.org/usage/writing-tests
 import React, { PropsWithChildren } from "react";
+import axios from "axios";
 import { Provider } from "react-redux";
 import { MemoryRouter } from "react-router-dom";
 import { render, screen, RenderOptions } from "@testing-library/react";
 import { UserEvent } from "@testing-library/user-event/dist/types/setup/setup";
 import userEvent from "@testing-library/user-event";
 import { PreloadedState } from "@reduxjs/toolkit";
-import { setupStore, AppStore } from "../redux/store";
-import { getResponseFromRecordName } from "./hooks/useInjectSearchResponse";
-import { RootState } from "../redux/store";
-import { FakeResponseName } from "./hooks/useInjectSearchResponse";
+import {
+  getResponseFromRecordName,
+  FakeResponseName,
+} from "./hooks/useInjectSearchResponse";
+import store, {
+  setupStore,
+  clearAllData,
+  AppStore,
+  RootState,
+} from "../redux/store";
 import { default as initialSearchState } from "../redux/search/initialState";
-import { initialState as initialSearchFormState } from "../redux/searchFormSlice";
+import {
+  initialState as initialSearchFormState,
+  setSearchFormDate,
+} from "../redux/searchFormSlice";
 import { processCharges } from "../redux/search/actions";
+import App from "../components/App";
 
 interface ExtendedRenderOptions extends Omit<RenderOptions, "queries"> {
   preloadedState?: PreloadedState<RootState>;
   store?: AppStore;
 }
 
+// setup helpers
 export function createStore(fakeResponseName?: FakeResponseName) {
   let search = initialSearchState;
 
@@ -64,6 +76,30 @@ export function setupUserAndRender(
 ) {
   const user = userEvent.setup();
   return { user, ...appRender(ui, fakeResponseName, opts) };
+}
+
+export function setupIntegrationTests(component = <App />) {
+  store.dispatch(clearAllData());
+  store.dispatch(setSearchFormDate("1/2/2023"));
+
+  const user = userEvent.setup();
+  const requestSpy = jest.spyOn(axios, "request").mockResolvedValue({});
+
+  appRender(component, undefined, {
+    store,
+  });
+  return { user, requestSpy };
+}
+
+// assertion helpers
+export function assertRequest(spy: jest.SpyInstance, obj: Object) {
+  expect(spy).toHaveBeenCalledWith(obj);
+  spy.mockClear();
+}
+
+// user helpers
+export async function goToSearchPage(user: UserEvent) {
+  await user.click(screen.getAllByRole("link", { name: /search/i })[0]);
 }
 
 type ButtonName =
@@ -152,8 +188,8 @@ export async function fillNewCaseForm(user: UserEvent) {
   await user.keyboard("1999");
 }
 
-export async function fillExpungementPacketForm(user: UserEvent) {
-  await user.click(screen.getByLabelText(/date of birth/i));
+export async function fillExpungementPacketForm(user: UserEvent, dobIndex = 0) {
+  await user.click(screen.getAllByLabelText(/birth/i)[dobIndex]);
   await user.keyboard("12/12/1999");
 
   await user.click(screen.getByLabelText(/address/i));
